hydrate Session into Server
